Refactored die_graph and implemented end_turn button.
Refactored die graph to use structs instead of objects, and added all functionality necessary for an end_turn button that removes block. Also added functionality that would allow for more general handling of tags. Including the ability for hit to be delayed to the end of the round, still needs visualisation and state logic before it can be finished.